Move crafting related stuff to new package
