+++ src/main/java/net/glowstone/inventory/GlowCraftingInventory.java
@@ -1,6 +1,7 @@
 package net.glowstone.inventory;
 
 import net.glowstone.GlowServer;
+import net.glowstone.inventory.crafting.CraftingManager;
 import org.bukkit.Bukkit;
 import org.bukkit.event.inventory.InventoryType;
 import org.bukkit.inventory.CraftingInventory;
